Use custom user manager for username normalization
Since Django 1.6, this can be used to customize user login with less impact than
subclassing the entire ModelBackend.

FWIW, our copy of the code was missing a security fix for an account
enumeration: https://code.djangoproject.com/ticket/20760
